+++ PixelTest/ImageType.swift
@@ -1,0 +1,14 @@
+//
+//  ImageType.swift
+//  PixelTest
+//
+//  Created by Kane Cheshire on 19/03/2018.
+//
+
+import Foundation
+
+enum ImageType: String {
+    case reference = "Reference"
+    case diff = "Diff"
+    case failure = "Failure"
+}
